+++ xbmc/music/Song.h
@@ -34,7 +34,7 @@ class CVariant;
 class CGenre
 {
 public:
-  long idGenre;
+  int idGenre;
   std::string strGenre;
 };
 
@@ -157,7 +157,7 @@ public:
   void SetArtistCredits(const std::vector<std::string>& names, const std::vector<std::string>& hints,
     const std::vector<std::string>& mbids);
 
-  long idSong;
+  int idSong;
   int idAlbum;
   std::string strFileName;
   std::string strTitle;
